Moving tests from Metadata.Tests to Datahub.Tests project
