+++ Datahub.CKAN/Datahub.CKAN.csproj
@@ -1,0 +1,12 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net6.0</TargetFramework>
+    <RootNamespace>NRCan.Datahub.CKAN</RootNamespace>    
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Datahub.Metadata\Datahub.Metadata.csproj" />
+  </ItemGroup>
+
+</Project>
