+++ Datahub.CKAN/Package/TranslatedFieldAgent.cs
@@ -1,7 +1,7 @@
 ﻿using NRCan.Datahub.Metadata.Model;
 using System.Collections.Generic;
 
-namespace NRCan.Datahub.Metadata.CKAN
+namespace NRCan.Datahub.CKAN.Package
 {
     class TranslatedFieldAgent : FieldAgent
     {
