+++ Datahub.CKAN/Service/CKANSeviceCollectionExtensions.cs
@@ -1,0 +1,12 @@
+﻿using Microsoft.Extensions.DependencyInjection;
+
+namespace NRCan.Datahub.CKAN.Service
+{
+    public static class CKANSeviceCollectionExtensions
+    {
+        public static void AddCKANService(this IServiceCollection services)
+        {
+            services.AddScoped<ICKANService, CKANService>();
+        }
+    }
+}
