+++ DatahubTest/Metadata/MetadataDefinitionsTests.cs
@@ -1,11 +1,10 @@
 using NRCan.Datahub.Metadata.DTO;
 using NRCan.Datahub.Metadata.Model;
 using System;
-using System.Collections.Generic;
 using System.Linq;
 using Xunit;
 
-namespace NRCan.Datahub.Metadata.Tests
+namespace DatahubTest
 {
     public class MetadataDefinitionsTests
     {
