+++ WebPortal/_Imports.razor
@@ -39,3 +39,4 @@
 @using Microsoft.Extensions.Configuration
 @using NRCan.Datahub.Shared.EFCore
 @using NRCan.Datahub.Portal.Components.Metadata;
+@using NRCan.Datahub.CKAN.Service;
